backend/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
billiard==4.2.0
blinker==1.8.2
Brotli==1.1.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
debugpy==1.8.1
decorator==4.4.2
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
Flask==3.0.3
Flask-Cors==4.0.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.1
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kombu==5.3.7
lz4==4.3.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
moviepy==1.0.3
mutagen==1.47.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.34.0
openpyxl==3.1.4
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pinecone==4.0.0
pinecone-client==5.0.0
pinecone-plugin-inference==1.0.2
pinecone-plugin-interface==0.0.7
platformdirs==4.2.2
proglog==0.1.10
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
protoc-gen-openapiv2==0.0.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodomex==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.3
redis==5.0.7
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
yt-dlp==2024.7.16
